Fix crash when sending position/SOS: JSONObject.put Float overload missing
BraceletApiClient.syncPosition and syncSos called put("accuracy", precision)
with a Float. Android's built-in org.json.JSONObject has no put(String, float)
overload, so every location sync and SOS threw NoSuchMethodError and killed the
TrackerService before the network call, leaving nothing sent to the backend.

Cast the Float to Double so it resolves to the existing put(String, double).

+++ app/src/main/java/com/camertrack/bracelet/data/BraceletApiClient.kt
@@ -115,7 +115,7 @@ object BraceletApiClient {
                 put("id", System.currentTimeMillis().toString())
                 put("latitude", latitude)
                 put("longitude", longitude)
-                put("accuracy", precision)
+                put("accuracy", precision.toDouble())
                 put("timestamp", java.time.Instant.now().toString())
             }
             val corps = "[$position]".toRequestBody(JSON)
@@ -157,7 +157,7 @@ object BraceletApiClient {
                 put("id", System.currentTimeMillis().toString())
                 put("latitude", latitude)
                 put("longitude", longitude)
-                put("accuracy", precision)
+                put("accuracy", precision.toDouble())
                 put("timestamp", java.time.Instant.now().toString())
                 put("severity", severity)
             }
